requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
PyYAML>=6.0.1
joblib>=1.3.0
scipy>=1.11.0
scikit-learn==1.6.1
mlflow==2.22.4
streamlit>=1.30.0
